+++ lua-api-crates/termwiz-funcs/src/lib.rs
@@ -9,6 +9,7 @@ use termwiz::color::{AnsiColor, ColorAttribute, ColorSpec, SrgbaTuple};
 use termwiz::input::Modifiers;
 use termwiz::render::terminfo::TerminfoRenderer;
 use termwiz::surface::change::Change;
+use termwiz::surface::Line;
 use wezterm_dynamic::{FromDynamic, ToDynamic};
 
 pub fn register(lua: &Lua) -> anyhow::Result<()> {
@@ -261,4 +262,41 @@ lazy_static::lazy_static! {
 
 pub fn new_wezterm_terminfo_renderer() -> TerminfoRenderer {
     TerminfoRenderer::new(CAPS.clone())
+}
+
+pub fn lines_to_escapes(lines: Vec<Line>) -> anyhow::Result<String> {
+    let mut changes = vec![];
+    let mut attr = CellAttributes::blank();
+    for line in lines {
+        changes.append(&mut line.changes(&attr));
+        changes.push(Change::Text("\r\n".to_string()));
+        if let Some(a) = line.visible_cells().last().map(|cell| cell.attrs().clone()) {
+            attr = a;
+        }
+    }
+    changes.push(Change::AllAttributes(CellAttributes::blank()));
+    let mut renderer = new_wezterm_terminfo_renderer();
+
+    struct Target {
+        target: Vec<u8>,
+    }
+
+    impl std::io::Write for Target {
+        fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
+            std::io::Write::write(&mut self.target, buf)
+        }
+        fn flush(&mut self) -> std::io::Result<()> {
+            Ok(())
+        }
+    }
+
+    impl termwiz::render::RenderTty for Target {
+        fn get_size_in_cells(&mut self) -> termwiz::Result<(usize, usize)> {
+            Ok((80, 24))
+        }
+    }
+
+    let mut target = Target { target: vec![] };
+    renderer.render_to(&changes, &mut target)?;
+    Ok(String::from_utf8(target.target)?)
 }