+++ wezterm-client/src/client.rs
@@ -1204,6 +1204,11 @@ impl Client {
         LivenessResponse
     );
     rpc!(get_lines, GetLines, GetLinesResponse);
+    rpc!(
+        get_dimensions,
+        GetPaneRenderableDimensions,
+        GetPaneRenderableDimensionsResponse
+    );
     rpc!(get_codec_version, GetCodecVersion, GetCodecVersionResponse);
     rpc!(get_tls_creds, GetTlsCreds = (), GetTlsCredsResponse);
     rpc!(